reorganize package structure, moving downloader code into separate modules
